requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
slowapi==0.1.9
pydantic[email]==2.5.0
email-validator==2.1.0
